Integrate GUI with search algorithms (JS)
